src/chickenrun/requirements.txt
paho-mqtt
jsonschema
gpiozero
